refactor: move createServer to service
